Creating tomcat_speech package, updating .gitignore, removing pyproject.toml
