+++ .gitignore
@@ -1,3 +1,6 @@
+# macOS
+.DS_Store
+
 # intellij
 .idea
