--- pyproject.toml
@@ -1,27 +1,0 @@
-[project]
-name = "tcspeech"
-version = "0.0.1"
-description = "ToMCAT Speech Analysis"
-homepage = "https://github.com/clulab/tomcat-speech"
-python = "^3.7"
-authors = [
-    "Rebecca Sharp",
-    "John Culnan",
-    "Meghavarshini Krishnaswamy",
-    "Adarsh Pyarelal"
-]
-readme = "README.md"
-keywords = ["speech analysis"]
-
-dependencies = [
-    "wheel",
-    "torch",
-    "pandas",
-    "numpy",
-    "sklearn",
-    "matplotlib"
-]
-
-[build-system]
-requires = ["setuptools>=41.0", "setuptools-scm", "wheel"]
-build-backend = "setuptools.build_meta"
