+++ tomcat_speech/data_prep/lives_data/lives_prep.py
@@ -11,7 +11,7 @@ import functools
 import operator
 import os
 
-from data_prep.data_prep_helpers import (
+from tomcat_speech.data_prep.data_prep_helpers import (
     MinMaxScaleRange,
     get_longest_utterance,
     clean_up_word,
