+++ tomcat_speech/models/parameters/multichannel_cnn_params.py
