+++ tomcat_speech/train_and_test_models/manipulate_lives_input.py
