+++ tomcat_speech/train_and_test_models/test_meld.py
@@ -3,15 +3,15 @@
 
 import pickle
 
-from data_prep.data_prep_helpers import DatumListDataset
-from models.train_and_test_models import *
+from tomcat_speech.data_prep.data_prep_helpers import DatumListDataset
+from tomcat_speech.models.train_and_test_models import *
 
-from models.input_models import *
-from data_prep.data_prep import *
-from data_prep.meld_data.meld_prep import *
+from tomcat_speech.models.input_models import *
+from tomcat_speech.data_prep.data_prep import *
+from tomcat_speech.data_prep.meld_data.meld_prep import *
 
 # import parameters for model
-from models.parameters.multitask_params import params
+from tomcat_speech.models.parameters.multitask_params import params
 
 # set device
 cuda = False
