+++ tomcat_speech/train_and_test_models/train_meld.py
@@ -1,4 +1,4 @@
-# train the models created in models directory with MUStARD data
+# train the models created in models directory with MELD data
 # currently the main entry point into the system
 
 import sys
@@ -6,16 +6,17 @@ import numpy as np
 
 from sklearn.model_selection import train_test_split
 
-from models.train_and_test_models import *
-from models.input_models import *
-from data_prep.data_prep_helpers import *
-from data_prep.meld_data.meld_prep import *
-from data_prep.mustard_data.mustard_prep import *
+sys.path.append("/net/kate/storage/work/bsharp/github/asist-speech")
+
+from tomcat_speech.models.train_and_test_models import *
+
+from tomcat_speech.models.input_models import *
+from tomcat_speech.data_prep.data_prep_helpers import *
+from tomcat_speech.data_prep.meld_data.meld_prep import *
+from tomcat_speech.data_prep.mustard_data.mustard_prep import *
 
 # import parameters for model
-from models.parameters.latefusion_params import params
-
-sys.path.append("/net/kate/storage/work/bsharp/github/asist-speech")
+from tomcat_speech.models.parameters.multitask_params import params
 
 # set device
 cuda = False
@@ -41,10 +42,13 @@ random.seed(seed)
 glove_file = "../../glove.short.300d.punct.txt"
 # glove_file = "../../glove.42B.300d.txt"
 
-mustard_path = "../../datasets/multimodal_datasets/MUStARD"
+# meld_path = "/data/nlp/corpora/MM/MELD_five_dialogues"
+# meld_path = "/data/nlp/corpora/MM/MELD_formatted"
+meld_path = "../../datasets/multimodal_datasets/MELD_formatted"
+# meld_path = "../../datasets/multimodal_datasets/MELD_five_utterances"
+# meld_path = "../../datasets/multimodal_datasets/MUStARD"
 
-data_type = "mustard"
-fusion_type = "late"
+data_type = "meld"
 
 # set model name and model type
 model = params.model
@@ -70,22 +74,37 @@ if __name__ == "__main__":
 
     # 2. MAKE DATASET
     # meld_data = MustardPrep(mustard_path=meld_path)
-    data = MustardPrep(
-        mustard_path=mustard_path,
+    # data = MeldPrep(
+    #     meld_path=meld_path,
+    #     acoustic_length=params.audio_dim,
+    #     glove=glove,
+    #     add_avging=params.add_avging,
+    #     avgd=avgd_acoustic,
+    # )
+
+    data = MeldPrep(
+        meld_path=meld_path,
         acoustic_length=params.audio_dim,
         glove=glove,
         add_avging=params.add_avging,
+        use_cols=[
+            "pcm_loudness_sma",
+            "F0finEnv_sma",
+            "voicingFinalUnclipped_sma",
+            "jitterLocal_sma",
+            "shimmerLocal_sma",
+            "pcm_loudness_sma_de",
+            "F0finEnv_sma_de",
+            "voicingFinalUnclipped_sma_de",
+            "jitterLocal_sma_de",
+            "shimmerLocal_sma_de",
+        ],
         avgd=avgd_acoustic,
     )
 
-    # data = MustardPrep(mustard_path=mustard_path, acoustic_length=params.audio_dim, add_avging=params.add_avging,
-    #                 use_cols=['pcm_loudness_sma', 'F0finEnv_sma', 'voicingFinalUnclipped_sma', 'jitterLocal_sma',
-    #                           'shimmerLocal_sma', 'pcm_loudness_sma_de', 'F0finEnv_sma_de',
-    #                           'voicingFinalUnclipped_sma_de', 'jitterLocal_sma_de', 'shimmerLocal_sma_de'],
-    #                 avgd=avgd_acoustic)
-
     # add class weights to device
-    data.sarcasm_weights = data.sarcasm_weights.to(device)
+    data.emotion_weights = data.emotion_weights.to(device)
+    data.sentiment_weights = data.sentiment_weights.to(device)
 
     print("Dataset created")
 
@@ -97,7 +116,6 @@ if __name__ == "__main__":
 
     # prepare holders for loss and accuracy of best model versions
     all_test_losses = []
-    all_test_accs = []
 
     # mini search through different learning_rate values
     for lr in params.lrs:
@@ -105,7 +123,11 @@ if __name__ == "__main__":
             # model_type = f"Multitask_1.6vs1lossWeighting_Adagrad_TextOnly_100batch_wd{str(wd)}_.2split"
             # model_type = f"TextOnly_smallerPool_100batch_wd{str(wd)}_.2split_500hidden"
             # model_type = f"AcousticGenderAvgd_noBatchNorm_.2splitTrainDev_IS10avgdAI_100batch_wd{str(wd)}_30each"
-            model_type = "MUStARD_lateFusionTest_avgdAcoustic_BothGendEmbs"
+            # model_type = "DELETE_ME_extraAudioFCs_.4drpt_Acou20Hid100Out"
+            model_type = (
+                "EMOTION_MODEL_FOR_ASIST"
+                # "MELD_IS10sm_500txthid_.1InDrpt_.3textdrpt_.4acdrpt_.5finalFCdrpt"
+            )
 
             # this uses train-dev-test folds
             # create instance of model
@@ -123,15 +145,6 @@ if __name__ == "__main__":
                 )
                 # optimizer = torch.optim.Adam(lr=lr, params=bimodal_trial.parameters(),
                 #                              weight_decay=wd)
-            elif fusion_type == "late":
-                bimodal_trial = LateFusionMultimodalModel(
-                    params=params,
-                    num_embeddings=num_embeddings,
-                    pretrained_embeddings=pretrained_embeddings,
-                )
-                optimizer = torch.optim.Adam(
-                    lr=lr, params=bimodal_trial.parameters(), weight_decay=wd
-                )
             elif params.text_only:
                 bimodal_trial = TextOnlyCNN(
                     params=params,
@@ -162,9 +175,10 @@ if __name__ == "__main__":
             print(bimodal_trial)
 
             # set loss function, optimization, and scheduler, if using
-            # loss_func = nn.CrossEntropyLoss(reduction="mean")
-            loss_func = nn.BCELoss(reduction="mean")
-            # loss_func = nn.CrossEntropyLoss(data.sarcasm_weights, reduction='mean')
+            loss_func = nn.CrossEntropyLoss(reduction="mean")
+            # loss_func = nn.CrossEntropyLoss(data.emotion_weights, reduction='mean')
+
+            # optimizer = torch.optim.SGD(bimodal_trial.parameters(), lr=lr, momentum=0.9)
 
             print("Model, loss function, and optimization created")
 
@@ -172,18 +186,28 @@ if __name__ == "__main__":
             # train_data = data.train_data
 
             # combine train and dev data
-            train_ds = DatumListDataset(
-                data.train_data, data_type, data.sarcasm_weights
-            )
-            dev_ds = DatumListDataset(data.dev_data, data_type, data.sarcasm_weights)
-            test_ds = DatumListDataset(data.test_data, data_type, data.sarcasm_weights)
+            train_and_dev = data.train_data + data.dev_data
 
-            train_targets = torch.stack(list(train_ds.targets()))
-            sampler_weights = data.sarcasm_weights
-            train_samples_weights = sampler_weights[train_targets]
-            sampler = torch.utils.data.sampler.WeightedRandomSampler(
-                train_samples_weights, len(train_samples_weights)
+            train_data, dev_data = train_test_split(train_and_dev, test_size=0.2)  # .3
+
+            train_ds = DatumListDataset(
+                train_data, data_type="meld_emotion", class_weights=data.emotion_weights
             )
+            # train_targets = torch.stack(list(train_ds.targets()))
+            # sampler_weights = data.emotion_weights
+            # train_samples_weights = sampler_weights[train_targets]
+            # sampler = torch.utils.data.sampler.WeightedRandomSampler(
+            #     train_samples_weights, len(train_samples_weights)
+            # )
+
+            dev_ds = DatumListDataset(dev_data, data.emotion_weights)
+            # dev_ds = DatumListDataset(data.dev_data, data.emotion_weights)
+            test_ds = DatumListDataset(data.test_data, data.emotion_weights)
+
+            #
+            # train_ds = data.train_data
+            # dev_ds = data.dev_data
+            # test_ds = data.test_data
 
             # create a a save path and file for the model
             model_save_file = "{0}_batch{1}_{2}hidden_2lyrs_lr{3}.pth".format(
@@ -192,9 +216,6 @@ if __name__ == "__main__":
 
             # make the train state to keep track of model training/development
             train_state = make_train_state(lr, model_save_path, model_save_file)
-
-            # set the load path for testing
-            load_path = model_save_path + model_save_file
 
             # train the model and evaluate on development set
             if multitask:
@@ -230,7 +251,6 @@ if __name__ == "__main__":
                     avgd_acoustic=avgd_acoustic_in_network,
                     use_speaker=params.use_speaker,
                     use_gender=params.use_gender,
-                    binary=True,
                 )
 
             # plot the loss and accuracy curves
